+++ apps/evm/src/containers/LiquidityHubForm/WithdrawForm/index.tsx
@@ -7,7 +7,6 @@ import { useTranslation } from 'libs/translations';
 import type { LiquidityHub, LiquidityHubBalanceMutation } from 'types';
 import { convertTokensToMantissa, formatTokensToReadableValue } from 'utilities';
 import { Form, type FormValues, initialFormValues } from '../Form';
-import { formatUserMaxTokenValue } from '../formatUserMaxTokenValue';
 
 export interface WithdrawFormProps {
   liquidityHub: LiquidityHub;
@@ -18,17 +17,16 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = ({ liquidityHub, onSubm
   const { t } = useTranslation();
   const [formValues, setFormValues] = useState(initialFormValues);
 
-  const userMaxRedeemTokens = formatUserMaxTokenValue({
-    value: liquidityHub.userVhTokenMaxRedeemTokens?.multipliedBy(liquidityHub.pricePerShare),
-    decimals: liquidityHub.vhToken.underlyingToken.decimals,
-  });
+  const { decimals } = liquidityHub.vhToken.underlyingToken;
 
-  // Apply buffer to account for accruing interests that lower the limits while a transaction is
-  // being executed
-  const userWithdrawCapTokens = formatUserMaxTokenValue({
-    value: liquidityHub.userWithdrawCapTokens,
-    decimals: liquidityHub.vhToken.underlyingToken.decimals,
-  });
+  const userMaxRedeemTokens = liquidityHub.userVhTokenMaxRedeemTokens
+    ?.multipliedBy(liquidityHub.pricePerShare)
+    .dp(decimals, BigNumber.ROUND_DOWN);
+
+  const userWithdrawCapTokens = liquidityHub.userWithdrawCapTokens?.dp(
+    decimals,
+    BigNumber.ROUND_DOWN,
+  );
 
   const limitTokens = BigNumber.min(userWithdrawCapTokens ?? 0, userMaxRedeemTokens ?? 0);
 
@@ -78,7 +76,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = ({ liquidityHub, onSubm
     ? () =>
         setFormValues(values => ({
           ...values,
-          amountTokens: limitTokens.dp(liquidityHub.vhToken.underlyingToken.decimals).toFixed(),
+          amountTokens: limitTokens.dp(decimals).toFixed(),
         }))
     : undefined;